+++ projects/layer-control/src/lib/layerentry/layerentry.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit, Input, Output, EventEmitter, HostBinding } from '@an
 
 // imports only for typings...
 import {
-  LayerGroup, Layer, RasterLayer, isRasterLayertype, WmsLayertype, WmtsLayertype, isRasterLayer,
-  isVectorLayer, LayersService
+  LayerGroup, Layer, RasterLayer, LayersService, WmsLayer, WmtsLayer
 } from '@dlr-eoc/services-layers';
 import { MapStateService } from '@dlr-eoc/services-map-state';
 
@@ -240,33 +239,27 @@ export class LayerentryComponent implements OnInit {
   }
 
   isSelectedStyle(styleName: string): boolean {
-    if (isRasterLayer(this.layer)) {
-      if (this.layer.type === WmsLayertype) {
-        return (this.layer as RasterLayer).params.STYLES === styleName;
-      } else if (this.layer.type === WmtsLayertype) {
-        return (this.layer as RasterLayer).params.style === styleName;
-      }
-    } else if (isVectorLayer(this.layer)) {
-      // TODO: how to compare styles for vector layers?
+    if (this.layer instanceof WmsLayer) {
+      return this.layer.params.STYLES === styleName;
+    } else if (this.layer instanceof WmtsLayer) {
+      return this.layer.params.style === styleName;
+    } else {
+      // TODO: how to compare styles for vector layers and custom layers?
       return false;
     }
-    // TODO: how to compare styles for custom layers?
-    return false;
   }
 
   executeChangeStyle(newStyleName: string) {
-    if (isRasterLayertype(this.layer.type)) {
-      if ((this.layer as RasterLayer).styles) {
-        const newStyle = (this.layer as RasterLayer).styles.find(s => s.name === newStyleName);
-        if (newStyle) {
-          this.layer.legendImg = newStyle.legendURL;
-          if (this.layer.type === WmsLayertype) {
-            (this.layer as RasterLayer).params.STYLES = newStyle.name;
-          } else if (this.layer.type === WmtsLayertype) {
-            (this.layer as RasterLayer).params.style = newStyle.name;
-          }
-          this.layersSvc.updateLayer(this.layer, this.layer.filtertype);
+    if (this.layer.styles) {
+      const newStyle = (this.layer as RasterLayer).styles.find(s => s.name === newStyleName);
+      if (newStyle) {
+        this.layer.legendImg = newStyle.legendURL;
+        if (this.layer instanceof WmsLayer) {
+          this.layer.params.STYLES = newStyle.name;
+        } else if (this.layer instanceof WmtsLayer) {
+          this.layer.params.style = newStyle.name;
         }
+        this.layersSvc.updateLayer(this.layer, this.layer.filtertype);
       }
     }
   }